.history/requirements_20251104094216.txt
# Backend and API
fastapi
uvicorn[standard]
pydantic
psycopg2-binary # PostgreSQL adapter
sqlalchemy # For ORM
alembic # For database migrations

# Asynchronous Tasks
celery
redis

# Document Processing (will add more later, starting with these)
python-docx
# install fitz/PyMuPDF separately (it has complex system dependencies, but often works as-is)
# We will use PyMuPDF or fitz later, for now, focus on the base
# If using PyMuPDF, it often requires system libs, but we'll try a simple install first:
PyMuPDF
pdfplumber
Pillow # For image manipulation before OCR

# AI/ML Libraries (Placeholder, will expand)
scikit-learn
numpy